+++ src/main/java/lotto/domain/LottoTicket.java
@@ -4,9 +4,9 @@ import camp.nextstep.edu.missionutils.Randoms;
 
 import java.util.List;
 
-import static lotto.constant.NumberConstant.LOTTO_LENGTH;
-import static lotto.constant.NumberConstant.LOTTO_MIN_VALUE;
-import static lotto.constant.NumberConstant.LOTTO_MAX_VALUE;
+import static lotto.common.NumberConstant.LOTTO_LENGTH;
+import static lotto.common.NumberConstant.LOTTO_MIN_VALUE;
+import static lotto.common.NumberConstant.LOTTO_MAX_VALUE;
 
 public class LottoTicket {
 
@@ -14,6 +14,5 @@ public class LottoTicket {
         List<Integer> lottoNumbers = Randoms.pickUniqueNumbersInRange(LOTTO_MIN_VALUE, LOTTO_MAX_VALUE, LOTTO_LENGTH);
         return new Lotto(lottoNumbers);
     }
-
 
 }
