+++ src/main/java/lotto/view/InputValidator.java
@@ -3,7 +3,7 @@ package lotto.view;
 import java.util.List;
 import java.util.regex.Pattern;
 
-import static lotto.constant.NumberConstant.LOTTO_LENGTH;
+import static lotto.common.NumberConstant.LOTTO_LENGTH;
 
 public class InputValidator {
