+++ src/main/java/lotto/view/InputView.java
@@ -9,7 +9,6 @@ import java.util.stream.Collectors;
 public class InputView {
     private static final String REGEX = ",";
     private static final String LINE = System.lineSeparator();
-
     private static final String INPUT_PRICE = "구입금액을 입력해 주세요.";
     private static final String INPUT_LOTTO_NUMBERS = LINE + "당첨 번호를 입력해 주세요.";
     private static final String INPUT_BONUS_NUMBER = LINE + "보너스 번호를 입력해 주세요.";
@@ -40,4 +39,5 @@ public class InputView {
                 .map(Integer::parseInt)
                 .collect(Collectors.toList());
     }
+
 }
